refactor out proxies into separate module
